+++ src/value.c
@@ -168,7 +168,7 @@ int val_compact_list( UCHAR *nodes_compact_list, UCHAR *target_id ) {
 	NODE_V *node_v = NULL;
 	TARGET_V *target = NULL;
 	ITEM *item = NULL;
-	ULONG j = 0;
+	int j = 0;
 	int size = 0;
 
 	/* Look into the local database */
